Refactor image loading and texture handling, update to A-Frame 1.7.1
Don't clear the texture cache all the time.
Load the image and when loaded transfer the texture and update the
material.

Added a test image to the assets as a default texture source.
Whith this the shader is initialized already with a texture.
Therefore we can avoid recompiling the shader later after adding the first
texture.